:sparles: feat: migrations

+++ database/migrations/2023_07_20_182028_create_reads_table.php
@@ -17,10 +17,10 @@ return new class extends Migration
             $table->foreign('id_book')->references('id')->on('books')->onUpdate('cascade')->onDelete('cascade');
 
             $table->timestamp('timestamp');
-            $table->integer('time_read');
+            $table->double('time_read');
             $table->integer('stopped_page');
             $table->integer('pages_read');
-            $table->integer('time_read_per_page');
+            $table->double('time_read_per_page');
 
             $table->string('comments')->nullable();
